Removed Mesocycle repo dependency from DayFactory, now using EntityManager for creating JPA relationships

+++ src/main/java/com/noslen/training_tracker/dto/day/request/CreateDayRequest.java
@@ -4,8 +4,8 @@ import java.time.Instant;
 
 public record CreateDayRequest(
         Long mesoId,
-        Long week,
-        Long position,
+        Integer week,
+        Integer position,
         Instant createdAt,
         Instant updatedAt,
         String unit,
